Adicionando metodo de buscar e inserir vendedores

+++ backend/src/main/java/com/wd/projeto01/entities/Vendas.java
@@ -16,7 +16,7 @@ public class Vendas {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private Integer id;
 	private Integer visitas;
 	private Integer qtdVendas;
 	private Double valor;
@@ -30,7 +30,7 @@ public class Vendas {
 	public Vendas() {}
 
 
-	public Vendas(Long id, Integer visitas, Integer qtdVendas, Double valor, LocalDate data) {
+	public Vendas(Integer id, Integer visitas, Integer qtdVendas, Double valor, LocalDate data) {
 		super();
 		this.id = id;
 		this.visitas = visitas;
@@ -40,12 +40,12 @@ public class Vendas {
 	}
 
 
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 
 
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
