+++ backend/src/main/java/com/wd/projeto01/repositories/VendasRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.wd.projeto01.entities.Vendas;
 
-public interface VendasRepository extends JpaRepository<Vendas, Long>{
+public interface VendasRepository extends JpaRepository<Vendas, Integer>{
 
 }
