+++ backend/src/main/java/com/wd/projeto01/repositories/VendedorRepository.java
@@ -4,6 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.wd.projeto01.entities.Vendedor;
 
-public interface VendedorRepository extends JpaRepository<Vendedor, Long>{
+public interface VendedorRepository extends JpaRepository<Vendedor, Integer>{
 
+	
 }
